+++ plugin-dev/Source/Sentry/Private/Interface/SentrySamplingContextInterface.h
@@ -4,6 +4,8 @@
 
 #include "CoreMinimal.h"
 
+#include "SentryVariant.h"
+
 class ISentryTransactionContext;
 
 class ISentrySamplingContext
@@ -12,5 +14,5 @@ public:
 	virtual ~ISentrySamplingContext() = default;
 
 	virtual TSharedPtr<ISentryTransactionContext> GetTransactionContext() const = 0;
-	virtual TMap<FString, FString> GetCustomSamplingContext() const = 0;
+	virtual TMap<FString, FSentryVariant> GetCustomSamplingContext() const = 0;
 };