+++ plugin-dev/Source/Sentry/Private/SentrySubsystem.cpp
@@ -553,7 +553,7 @@ USentryTransaction* USentrySubsystem::StartTransactionWithContextAndTimestamp(US
 	return USentryTransaction::Create(SentryTransaction);
 }
 
-USentryTransaction* USentrySubsystem::StartTransactionWithContextAndOptions(USentryTransactionContext* Context, const TMap<FString, FString>& Options)
+USentryTransaction* USentrySubsystem::StartTransactionWithContextAndOptions(USentryTransactionContext* Context, const FSentryTransactionOptions& Options)
 {
 	check(SubsystemNativeImpl);
 	check(Context);
@@ -597,6 +597,11 @@ bool USentrySubsystem::IsSupportedForCurrentSettings() const
 	}
 
 	return true;
+}
+
+TSharedPtr<ISentrySubsystem> USentrySubsystem::GetNativeObject() const
+{
+	return SubsystemNativeImpl;
 }
 
 void USentrySubsystem::AddDefaultContext()